requirements.txt
Flask
Werkzeug
Jinja2
itsdangerous
click
pika
google-generativeai
retry
python-dotenv